LayoutTests/http/tests/websocket/tests/hybi/error-event-ready-state-non-existent-url-with-server-responding-404-expected.txt

Test that readyState is CLOSED within onerror event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onerror() was called.
PASS ws.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
